casting distance always to float

+++ src/main/scala/ch/unibas/dmi/dbis/adam/index/structures/lsh/LSHIndex.scala
@@ -65,7 +65,7 @@ class LSHIndex(val indexname: IndexName, val entityname: EntityName, override pr
           handler.offer(item, this.pk.name)
         })
 
-        handler.results.map(x => Row(x.tid, x.score)).iterator
+        handler.results.map(x => Row(x.tid, x.score.toFloat)).iterator
       }
 
     ac.sqlContext.createDataFrame(rddResults, Result.resultSchema(pk.name))
