+++ src/main/scala/ch/unibas/dmi/dbis/adam/index/structures/sh/SHIndex.scala
@@ -64,7 +64,7 @@ class SHIndex(val indexname: IndexName, val entityname: EntityName, override pri
           handler.offer(item, this.pk.name)
         })
 
-        handler.results.map(x => Row(x.tid, x.score)).iterator
+        handler.results.map(x => Row(x.tid, x.score.toFloat)).iterator
       }
 
     ac.sqlContext.createDataFrame(rddResults,  Result.resultSchema(pk.name))
